# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B13 CORINGA FFH4X API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type KeyDuration = typeof KeyDuration[keyof typeof KeyDuration];


export const KeyDuration = {
  '1min': '1min',
  '1day': '1day',
  '3days': '3days',
  '7days': '7days',
  '30days': '30days',
  permanent: 'permanent',
} as const;

export interface Key {
  id: number;
  keyValue: string;
  duration: KeyDuration;
  /** @nullable */
  expiresAt: string | null;
  active: boolean;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export type KeyGenerateInputDuration = typeof KeyGenerateInputDuration[keyof typeof KeyGenerateInputDuration];


export const KeyGenerateInputDuration = {
  '1min': '1min',
  '1day': '1day',
  '3days': '3days',
  '7days': '7days',
  '30days': '30days',
  permanent: 'permanent',
} as const;

export interface KeyGenerateInput {
  duration: KeyGenerateInputDuration;
  note?: string;
}

export interface KeyValidation {
  valid: boolean;
  key?: Key;
  message?: string;
}

export interface KeyStats {
  total: number;
  active: number;
  expired: number;
  permanent?: number;
}