# lib/api-zod/src/generated/types/key.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B13 CORINGA FFH4X API
 * OpenAPI spec version: 0.1.0
 */
import type { KeyDuration } from './keyDuration';

export interface Key {
  id: number;
  keyValue: string;
  duration: KeyDuration;
  /** @nullable */
  expiresAt: string | null;
  active: boolean;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}