# lib/api-zod/src/generated/types/keyGenerateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B13 CORINGA FFH4X API
 * OpenAPI spec version: 0.1.0
 */
import type { KeyGenerateInputDuration } from './keyGenerateInputDuration';

export interface KeyGenerateInput {
  duration: KeyGenerateInputDuration;
  note?: string;
}