# lib/api-zod/src/generated/types/keyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B13 CORINGA FFH4X API
 * OpenAPI spec version: 0.1.0
 */

export interface KeyStats {
  total: number;
  active: number;
  expired: number;
  permanent?: number;
}