# lib/api-zod/src/generated/types/keyValidation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B13 CORINGA FFH4X API
 * OpenAPI spec version: 0.1.0
 */
import type { Key } from './key';

export interface KeyValidation {
  valid: boolean;
  key?: Key;
  message?: string;
}